+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG for PHP CS Fixer: custom fixers
 
+## v3.28.0
+- Deprecate MultilinePromotedPropertiesFixer - use "multiline_promoted_properties"
+- Update minimum PHP CS Fixer version to 3.76.1
+
 ## v3.27.0
 - Add PhpdocTagNoNamedArgumentsFixer